requirements.txt
langchain
langchain-openai
openai
streamlit
faiss-cpu
python-dotenv
PyPDF2
